refactor(logger): correction of review edits
Renamed the logger to LoggerMiddleware. Also moved
to the middleware file.
